Release 1 1 12 (#153)
* [NR-439351]fix: Update agent versions and fix Android recordMetric

This commit updates the native Android agent to version 7.6.7 and the native iOS agent to version 7.5.6.
It also addresses an issue in the Android agent where the `recordMetric` method did not correctly handle null values for optional parameters. The Flutter agent version is updated to 1.1.12.

* fix: release flutter agent 1.1.13

This commit updates the native Android agent to version 7.6.8 and the native iOS agent to version 7.5.8. The Flutter agent version is updated to 1.1.13.